Add option for setting model root
